関数がintで宣言されているため小数点以下が切り落とされている

+++ main.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
-int myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。
+double myRoot(double x) // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。
 {
-  int y;
+  double y;
   if( x < 0 ){
     exit(1);
   }
@@ -17,7 +17,7 @@ int main(void)
   double *x = (double *)malloc(sizeof(double) * 100);
 
   while( i <= 100 ){
-    x[i] = myRoot((double)i);
+    x[i] = myRoot((int)i);
     i++;
   }
